requirements.txt
apipkg==1.5
asn1crypto==1.2.0
atomicwrites==1.3.0
attrs==19.3.0
backcall==0.1.0
baidu-api==0.0.2
beautifulsoup4==4.8.1
boto==2.49.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
colorama==0.4.1
configparser==4.0.2
coverage==4.5.4
cryptography==2.8
ddt==1.2.2
decorator==4.4.1
entrypoints==0.3
execnet==1.7.1
flake8==3.7.9
idna==2.8
importlib-metadata==1.1.0
ipdb==0.12.2
ipython==7.10.1
ipython-genutils==0.2.0
jedi==0.15.1
mccabe==0.6.1
module==0.0.4
more-itertools==8.0.0
nose==1.3.7
packaging==19.2
parameterized==0.7.1
parso==0.5.1
pickleshare==0.7.5
Pillow==7.0.0
pluggy==0.13.1
portalocker==1.5.2
prompt-toolkit==3.0.2
py==1.8.0
pycodestyle==2.5.0
pycparser==2.19
pycryptodome==3.9.4
pyflakes==2.1.1
Pygments==2.5.2
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyotp==2.3.0
pyparsing==2.4.5
pytest==5.3.2
pytest-cov==2.8.1
pytest-forked==1.1.3
pytest-html==2.0.1
pytest-metadata==1.8.0
pytest-ordering==0.6
pytest-rerunfailures==8.0
pytest-timeout==1.3.3
pytest-xdist==1.30.0
python-dateutil==2.8.1
pywin32==227
redis==3.3.11
requests==2.22.0
selenium==3.141.0
seleniumbase==1.33.9
setuptools-scm==3.3.3
six==1.13.0
sortedcontainers==2.1.0
soupsieve==1.9.5
tqdm==4.40.0
traitlets==4.3.3
urllib3==1.25.6
wcwidth==0.1.7
xlrd==1.2.0
zipp==0.6.0
